Move create_state implementation to source
